requirements.txt
flake8
icalendar
openpyxl
pyyaml
titlecase
Jinja2
